requirements.txt
anyio==4.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-scheduler==0.10.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
icalendar==5.0.12
idna==3.6
python-dateutil==2.9.0.post0
python-telegram-bot==21.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
